uWSGI/ScrapyNews/requirements.txt
django==2.1.5
pymysql
requests
bs4
djangorestframework==3.8.1
elasticsearch
scrapy
mysqlclient
scrapy_djangoitem
channels
channels_redis==2.3.3
confluent_kafka
pandas
uwsgi
django-cors-headers
